chore: merge payments into project boost (#1580)
* chore: merge new payments in project boosts

* chore: remove old payment components

* chore: rename file

* chore: styled component lint

* fix: amount input for project boost
