+++ packages/components/src/Payments/Payments.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { ICivilContext, CivilContext } from "../context";
 import { PaymentsAmount } from "./PaymentsAmount";
 import { PaymentsLoginOrGuest } from "./PaymentsLoginOrGuest";
-import { PaymentsOptions } from "./PaymentsOptions";
+import { PaymentsSelectType } from "./PaymentsSelectType";
 import { PaymentsEth } from "./PaymentsEth";
 import { PaymentsStripe } from "./PaymentsStripe";
 import { PaymentsApplePay } from "./PaymentsApplePay";
@@ -13,9 +13,12 @@ import { PaymentsSuccess } from "./PaymentsSuccess";
 
 export interface PaymentsProps {
   postId: string;
+  usdToSpend?: number;
+  boostType?: string;
   paymentAddress: string;
   newsroomName: string;
   isStripeConnected: boolean;
+  stripeAccountID: string;
   handleClose(): void;
 }
 
@@ -41,7 +44,7 @@ export class Payments extends React.Component<PaymentsProps, PaymentsStates> {
   constructor(props: any) {
     super(props);
     this.state = {
-      usdToSpend: 0,
+      usdToSpend: props.usdToSpend || 0,
       paymentAdjustedWarning: false,
       paymentAdjustedStripe: false,
       paymentAdjustedEth: false,
@@ -56,6 +59,14 @@ export class Payments extends React.Component<PaymentsProps, PaymentsStates> {
 
   public componentDidMount(): void {
     this.context.auth.ensureLoggedInUserEnabled();
+
+    if (this.props.boostType === "project") {
+      const paymentAdjustedWarning = this.props.usdToSpend && this.props.usdToSpend <= 2 ? true : false;
+      this.setState({
+        paymentState: PAYMENT_STATE.PAYMENT_CHOOSE_LOGIN_OR_GUEST,
+        paymentAdjustedWarning,
+      });
+    }
   }
 
   public render(): JSX.Element {
@@ -76,7 +87,15 @@ export class Payments extends React.Component<PaymentsProps, PaymentsStates> {
       paymentInProgress,
       waitingForConfirmation,
     } = this.state;
-    const { postId, paymentAddress, newsroomName, isStripeConnected, handleClose } = this.props;
+    const {
+      postId,
+      paymentAddress,
+      newsroomName,
+      isStripeConnected,
+      stripeAccountID,
+      boostType,
+      handleClose,
+    } = this.props;
     const showWeb3Login = this.context.auth.showWeb3Login;
 
     // User logged in from PAYMENT_CHOOSE_LOGIN_OR_GUEST state, which will be reflected in context, and we should now show them SELECT_PAYMENT_TYPE state instead.
@@ -95,13 +114,14 @@ export class Payments extends React.Component<PaymentsProps, PaymentsStates> {
     if (proceedToPaymentType || paymentState === PAYMENT_STATE.SELECT_PAYMENT_TYPE) {
       return (
         <PaymentsWrapper
+          boostType={boostType}
           usdToSpend={usdToSpend}
           newsroomName={newsroomName}
           paymentAdjustedWarning={paymentAdjustedWarning}
           handleEditAmount={() => this.handleUpdateState(PAYMENT_STATE.SELECT_AMOUNT)}
           handleBack={() => this.handleUpdateState(PAYMENT_STATE.SELECT_AMOUNT)}
         >
-          <PaymentsOptions
+          <PaymentsSelectType
             postId={postId}
             usdToSpend={usdToSpend}
             isStripeConnected={isStripeConnected}
@@ -117,6 +137,7 @@ export class Payments extends React.Component<PaymentsProps, PaymentsStates> {
     if (paymentState === PAYMENT_STATE.ETH_PAYMENT) {
       return (
         <PaymentsWrapper
+          boostType={boostType}
           usdToSpend={usdToSpend}
           newsroomName={newsroomName}
           paymentAdjustedEth={paymentAdjustedEth}
@@ -147,6 +168,7 @@ export class Payments extends React.Component<PaymentsProps, PaymentsStates> {
     if (paymentState === PAYMENT_STATE.STRIPE_PAYMENT) {
       return (
         <PaymentsWrapper
+          boostType={boostType}
           usdToSpend={usdToSpend}
           newsroomName={newsroomName}
           paymentAdjustedStripe={paymentAdjustedStripe}
@@ -158,6 +180,7 @@ export class Payments extends React.Component<PaymentsProps, PaymentsStates> {
             newsroomName={newsroomName}
             shouldPublicize={shouldPublicize}
             usdToSpend={usdToSpend}
+            stripeAccountID={stripeAccountID}
             handlePaymentSuccess={this.handlePaymentSuccess}
             handleEditPaymentType={() => this.handleUpdateState(PAYMENT_STATE.SELECT_PAYMENT_TYPE)}
           />
@@ -168,6 +191,7 @@ export class Payments extends React.Component<PaymentsProps, PaymentsStates> {
     if (paymentState === PAYMENT_STATE.APPLE_PAY) {
       return (
         <PaymentsWrapper
+          boostType={boostType}
           usdToSpend={usdToSpend}
           newsroomName={newsroomName}
           handleBack={() => this.handleUpdateState(PAYMENT_STATE.SELECT_PAYMENT_TYPE)}
@@ -184,6 +208,7 @@ export class Payments extends React.Component<PaymentsProps, PaymentsStates> {
     if (paymentState === PAYMENT_STATE.GOOGLE_PAY) {
       return (
         <PaymentsWrapper
+          boostType={boostType}
           usdToSpend={usdToSpend}
           newsroomName={newsroomName}
           handleBack={() => this.handleUpdateState(PAYMENT_STATE.SELECT_PAYMENT_TYPE)}
@@ -202,7 +227,7 @@ export class Payments extends React.Component<PaymentsProps, PaymentsStates> {
       paymentState === PAYMENT_STATE.PAYMENT_SUCCESS_WITH_SAVED_EMAIL
     ) {
       return (
-        <PaymentsWrapper newsroomName={newsroomName}>
+        <PaymentsWrapper newsroomName={newsroomName} boostType={boostType}>
           <PaymentsSuccess
             newsroomName={newsroomName}
             usdToSpend={usdToSpend}
@@ -217,7 +242,7 @@ export class Payments extends React.Component<PaymentsProps, PaymentsStates> {
     }
 
     return (
-      <PaymentsWrapper newsroomName={newsroomName} handleBack={handleClose}>
+      <PaymentsWrapper newsroomName={newsroomName} handleBack={handleClose} boostType={boostType}>
         <PaymentsAmount
           newsroomName={newsroomName}
           suggestedAmounts={SuggestedPaymentAmounts}
